chore: organized some components into folders and refactored to make use of notifications component
