+++ components/directory/CreateUserForm.js
@@ -130,7 +130,7 @@ export const CreateUserForm = ({
                 : "inline-flex w-full justify-center rounded-md bg-cyan-600 px-3 py-2 text-sm font-semibold text-white shadow-sm hover:bg-cyan-500 sm:ml-3 sm:w-auto"
             }
           >
-            Actualizar
+            Crear usuario
           </button>
           <button
             type="button"
