+++ components/directory/UserFormWrapper.js
@@ -1,7 +1,7 @@
 import { Divider } from "@/components/Divider";
 import { PersonalUserInfo } from "@/components/directory/PersonalUserInfo";
 import { UserNotes } from "@/components/directory/UserNotes";
-import { useNewUserForm } from "@/hooks/useNewUserForm";
+import { useNewUserForm } from "@/hooks/user_directory/useNewUserForm";
 //
 export const UserFormWrapper = () => {
   const { register, handleSubmit, errors, onSubmit, isSubmitting } =
@@ -21,22 +21,22 @@ export const UserFormWrapper = () => {
             </p>
           </div>
           <div className="py-12 px-2 md:px-0">
-            <div className="max-w-7xl mx-auto">
+            <div className="mx-auto max-w-7xl">
               <form onSubmit={handleSubmit(onSubmit)}>
                 <Divider />
                 <PersonalUserInfo register={register} errors={errors} />
                 <Divider />
                 <UserNotes register={register} />
 
-                <div className="pt-12 px-2">
+                <div className="px-2 pt-12">
                   <div className="flex justify-end">
                     <button
                       disabled={isSubmitting ?? false}
                       type="submit"
                       className={
                         isSubmitting
-                          ? "ml-3 inline-flex justify-center py-2 px-4 border border-transparent shadow-sm text-sm font-medium rounded-md text-gray-400 bg-gray-300"
-                          : "ml-3 inline-flex justify-center py-2 px-4 border border-transparent shadow-sm text-sm font-medium rounded-md text-white bg-cyan-700 hover:bg-cyan-500 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-cyan-500"
+                          ? "ml-3 inline-flex justify-center rounded-md border border-transparent bg-gray-300 py-2 px-4 text-sm font-medium text-gray-400 shadow-sm"
+                          : "ml-3 inline-flex justify-center rounded-md border border-transparent bg-cyan-700 py-2 px-4 text-sm font-medium text-white shadow-sm hover:bg-cyan-500 focus:outline-none focus:ring-2 focus:ring-cyan-500 focus:ring-offset-2"
                       }
                     >
                       Guardar
