+++ lib/user_directory/createFirebaseUser.js
