+++ lib/user_directory/deleteUser.js
