+++ lib/user_directory/getFirebaseUser.js
