+++ lib/user_directory/mergeUser.js
