+++ lib/user_directory/normalizeFirebaseUser.js
