+++ lib/user_directory/updateFirebaseUser.js
