+++ pages/organization-profile/[[...index]].js
@@ -1,4 +1,4 @@
-import { Container } from "@/components/container";
+import { Container } from "@/components/Container";
 import { OrganizationProfile } from "@clerk/nextjs";
 
 //
